Fix race in fsm.db
We need to take out a write lock when modifying any of the fields of the
FSM struct. We also need a read lock when reading any of the FSM fields.
Expose a new fsm.Stats to handle the read lock wrapping and make sure we
lock durring the f.db assignment.